Handle Qt.createComponent for compiled files
Add methods to QQmlEngine which allows one to set a callback that gets called in
place of creating a new QQmlComponent when Qt.createComponent is called from
javascript. We need this so qmlc loader can load the .qmc in place of it
creating new QQmlComponent for the .qml with Qt.createComponent calls.